+++ src/hooks/useEscapeKey.tsx
@@ -1,0 +1,21 @@
+import { useCallback, useEffect } from 'react';
+
+const useEscapeKey = (handleClose: () => void) => {
+  const handleKeyDown = useCallback(
+    (event: KeyboardEvent) => {
+      if (event.key === 'Escape') {
+        handleClose();
+      }
+    },
+    [handleClose]
+  );
+
+  useEffect(() => {
+    document.addEventListener('keydown', handleKeyDown);
+    return () => {
+      document.removeEventListener('keydown', handleKeyDown);
+    };
+  }, [handleKeyDown]);
+};
+
+export default useEscapeKey;
